fix: type imports, environment variables
