+++ src/lib/notion.ts
@@ -1,6 +1,6 @@
 import { Client, LogLevel } from '@notionhq/client';
 
 export const notion = new Client({
-  auth: process.env.NOTION_API_KEY,
-  logLevel: process.env.NODE_ENV === 'development' ? LogLevel.DEBUG : undefined,
+  auth: import.meta.env.NOTION_API_KEY,
+  logLevel: import.meta.env.NODE_ENV === 'development' ? LogLevel.DEBUG : undefined,
 });
