+++ src/lib/trpc.ts
@@ -2,9 +2,9 @@ import { createTRPCProxyClient, httpBatchLink } from '@trpc/client';
 import { createTRPCReact } from '@trpc/react-query';
 import type { AppRouter } from '~/server/routers/_app';
 
-const isNetlifyProdDeploy = process.env.CONTEXT === 'production';
+const isNetlifyProdDeploy = import.meta.env.CONTEXT === 'production';
 const isNetlifyPreviewOrBranchDeploy =
-  process.env.CONTEXT === 'deploy-preview' || process.env.CONTEXT === 'branch-deploy';
+  import.meta.env.CONTEXT === 'deploy-preview' || import.meta.env.CONTEXT === 'branch-deploy';
 
 function getBaseUrl() {
   if (typeof window !== 'undefined') {
@@ -12,28 +12,28 @@ function getBaseUrl() {
     return '';
   }
 
-  if (process.env.URL && isNetlifyProdDeploy) {
+  if (import.meta.env.URL && isNetlifyProdDeploy) {
     // reference for Netlify (production)
-    return process.env.DEPLOY_PRIME_URL;
+    return import.meta.env.DEPLOY_PRIME_URL;
   }
 
-  if (process.env.DEPLOY_PRIME_URL && isNetlifyPreviewOrBranchDeploy) {
+  if (import.meta.env.DEPLOY_PRIME_URL && isNetlifyPreviewOrBranchDeploy) {
     // reference for Netlify (deploy preview)
-    return process.env.DEPLOY_PRIME_URL;
+    return import.meta.env.DEPLOY_PRIME_URL;
   }
 
-  if (process.env.VERCEL_URL) {
+  if (import.meta.env.VERCEL_URL) {
     // reference for vercel.com
-    return `https://${process.env.VERCEL_URL}`;
+    return `https://${import.meta.env.VERCEL_URL}`;
   }
 
-  if (process.env.RENDER_INTERNAL_HOSTNAME) {
+  if (import.meta.env.RENDER_INTERNAL_HOSTNAME) {
     // reference for render.com
-    return `http://${process.env.RENDER_INTERNAL_HOSTNAME}:${process.env.PORT}`;
+    return `http://${import.meta.env.RENDER_INTERNAL_HOSTNAME}:${import.meta.env.PORT}`;
   }
 
   // assume localhost
-  return `http://localhost:${process.env.PORT ?? 3000}`;
+  return `http://localhost:${import.meta.env.PORT ?? 3000}`;
 }
 
 export const httpBatchLinkUrl = `${getBaseUrl()}/api/trpc`;
